+++ src/components/Poptip/Poptip.story.js
@@ -88,4 +88,15 @@ stories
         Customizable Poptip
       </Button>
     </Poptip>
+  )))
+  .add('basic poptip with delay', withInfo()(() => (
+    <Poptip
+      trigger='mouseenter'
+      delay={ 1000 }
+      hideDelay={ 5000 }
+      content="Ipsa officiis bad-news minus earum a aperiam! Aperiam reiciendis vitae nihil libero et, hic!">
+      <Button width="default">
+        Default Poptip
+      </Button>
+    </Poptip>
   )));